Fix Phase 2 compilation errors
Fixed all compilation errors to make Phase 2 self-sufficient:

1. Fixed TradeFilters initialization
   - Removed non-existent field `min_profit_per_volume`
   - Added required field `ship_id: None`

2. Fixed TradeOffer field references
   - Changed `profit` to `total_profit`
   - Changed `amount` to `qty`
   - Used `partial_cmp` for f64 sorting

3. Fixed type mismatches in DashboardRoute model
   - Changed `profit: i64` to `profit: f64`
   - Changed `buy_price: i32` to `buy_price: f64`
   - Changed `sell_price: i32` to `sell_price: f64`

4. Removed unused imports
   - Removed `std::collections::HashMap` from dashboard.rs
   - Removed `serde::Serialize` from pilot.rs
   - Removed unused exports from handlers/mod.rs

Phase 2 is now fully compilable and self-sufficient.

+++ src/models/dashboard.rs
@@ -33,9 +33,9 @@ pub struct DashboardRoute {
     pub buy_sector: String,
     pub sell_station: String,
     pub sell_sector: String,
-    pub profit: i64,
-    pub buy_price: i32,
-    pub sell_price: i32,
+    pub profit: f64,
+    pub buy_price: f64,
+    pub sell_price: f64,
     pub amount: i32,
 }
